# OpenAPI.yaml
openapi: 3.0.4
info:
  title: OpenAPI definition
  version: v1
servers:
  - url: http://localhost:8082
    description: Local server url
tags:
  - name: clock-in-rest-adapter
paths:
  /api/v1/clockin:
    get:
      tags:
        - clock-in-rest-adapter
      operationId: getClockInEntries
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClockInEntryResponse"
    post:
      tags:
        - clock-in-rest-adapter
      operationId: createClockInEntry
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockInEntryRequest"
        required: true
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/clockin/types/{id}:
    put:
      tags:
        - clock-in-rest-adapter
      operationId: updateClockInType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockInTypeRequest"
        required: true
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - clock-in-rest-adapter
      operationId: deleteClockInType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/v1/clockin/types:
    get:
      tags:
        - clock-in-rest-adapter
      operationId: getClockInTypes
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClockInTypeResponse"
    post:
      tags:
        - clock-in-rest-adapter
      operationId: createClockInType
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockInTypeRequest"
        required: true
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    ClockInEntryRequest:
      type: object
      properties:
        clockInType:
          type: string
          minLength: 1
        device:
          type: string
          minLength: 1
        comment:
          type: string
      required:
        - clockInType
        - device
    ClockInEntryResponse:
      type: object
      properties:
        id:
          type: string
        clockInType:
          $ref: "#/components/schemas/ClockInTypeResponse"
        date:
          type: string
          format: date
        time:
          type: string
        device:
          type: string
        comment:
          type: string
    ClockInTypeRequest:
      type: object
      properties:
        description:
          type: string
          minLength: 1
        io:
          type: boolean
      required:
        - description
        - io
    ClockInTypeResponse:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        io:
          type: boolean
    ApiResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string